+++ log-all-the-things/common/src/main/java/com/chimericdream/logallthethings/LogAllTheThingsMod.java
@@ -10,6 +10,8 @@ import org.apache.logging.log4j.Logger;
 
 import java.util.function.Supplier;
 
+import com.chimericdream.logallthethings.carpetlog.CarpetLogBlocks;
+import com.chimericdream.logallthethings.carpetlog.CarpetLogHelper;
 import com.chimericdream.logallthethings.windowlog.WindowLogBlocks;
 import com.chimericdream.logallthethings.windowlog.WindowLogHelper;
 
@@ -23,10 +25,13 @@ public final class LogAllTheThingsMod {
         MANAGER = Suppliers.memoize(() -> RegistrarManager.get(ModInfo.MOD_ID));
 
         WindowLogBlocks.init();
+        CarpetLogBlocks.init();
 
         REGISTRY_HELPER.init();
 
         InteractionEvent.RIGHT_CLICK_BLOCK.register(WindowLogHelper::tryWindowLog);
+        InteractionEvent.RIGHT_CLICK_BLOCK.register(CarpetLogHelper::tryCarpetLog);
         BlockEvent.BREAK.register(WindowLogHelper::tryPartialBreak);
+        BlockEvent.BREAK.register(CarpetLogHelper::tryPartialBreak);
     }
 }